Moved the content in form_pydantic.py to form.py
